Renamed Registration component and path in App
