+++ client/components/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import {Route} from 'react-router-dom'
+import { Route } from 'react-router-dom'
 import DropdownModule from './dropdownModule/DropdownModule'
 import Login from './Login'
-import RegistrationForm from './ResgistrationForm'
+import RegistrationForm from './RegistrationForm'
 import signin from './signin'
 import Dashboard from './Dashboard'
 
@@ -14,8 +14,8 @@ const App = () => {
 
       <h1>[React development has begun!]</h1>
       <Route Login path="/" component={Login} />
-      <Route path="/register" component={RegistrationForm}/>
-      < DropdownModule />
+      <Route path="/register" component={RegistrationForm} />
+      <DropdownModule />
     </React.Fragment>
   )
 }
